address review comments: css modules, component extraction, bug fixes
